requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
pyyaml>=6.0

# Optional: Advanced Models
# xgboost>=1.5.0
# lightgbm>=3.3.0

# Optional: Data Balancing
# imbalanced-learn>=0.9.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
